backend/requirements.txt
Flask==2.2.5
Flask_SQLAlchemy==3.0.3
Flask-JWT-Extended==4.4.4
Flask-Cors==4.0.1
python-dotenv==1.0.0
bcrypt==4.0.1
